Use macros instead of traits to extend models functionality
This commit add two macros to bring the models access to the methods filter and filterAndGet to the Builder class, inside the FilterServiceProvider. This way is no longer needed to add the trait to the model class. The name of the macros is also configurable, which by default are named the same as the class. This helps to avoid macros name collision in case there are already Builder macros in the application with the same names.