Big rework: Parsers and QueryBuilder classes, custom exceptions and config modifications
This commit makes a big rework to the structure of the library. For splitting the functionality, two groups of classes have been created: Parsers and QueryBuilders:

- Parsers are in charge of taking the key of each input and get the different parts of the query: the relationship to operate on, the column and the modifier (like, not, greater, etc...). Those parsers are used by the ColumnFilters class to build the constraints on the query.
- QueryBuilder classes add the constrains to the query: relationships to include, columns with constrains (included constrains on related models) and result sorting. The QueryBuilder/ColumnFilters class use the parsers to obtain each desired constrain from the keys of the input.

This way the responsabilities are more distributed. Also it's possible to test each responsability more accurately.

Custom exceptions were also added to be more precise about exceptions ocurred while using the library. A base exception for the library was included, so all the exceptions related to the library have all the same base class.

Some updates to the configuration has been included as well:

- The config keys relationship_separator and column_query_separator have been replaced by separators.relationship and separators.query_modifier respectively
- The paginate_by_default and page_size config entries have been removed. In order to paginate a query, the page_size parameter must be passed to the filter methods.